% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifair_base.R
\name{multifair_base}
\alias{multifair_base}
\title{Main solver of \code{HDfair}}
\usage{
multifair_base(
  x,
  y,
  group,
  lambda,
  eta,
  stepsize,
  intercept = FALSE,
  theta_init = NULL,
  type = "continuous",
  reg = "group-lasso",
  crit = "metric",
  rho = 1,
  maxit = 1000,
  tol = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{group}{Character string or integer indicating the name of the grouping variable or the column index of the grouping variable.}

\item{data}{A list containing the datasets from \eqn{M} sources.}

\item{outcome}{Character string or integer indicating the name of the outcome variable or the column index of the outcome variable.}
}
\description{
Main solver of \code{HDfair}
}
