% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDfair_cv_lambda.R
\name{plot_cv_lambda}
\alias{plot_cv_lambda}
\title{plot_cv_lambda: Plot Cross-Validation Error vs Lambda}
\usage{
plot_cv_lambda(cv_lambda)
}
\arguments{
\item{cv_lambda}{Object of class \code{"HDfair_cv"} returned by \code{HDfair_cv_lambda}.}
}
\value{
NULL. Generates a plot on the current graphics device.
}
\description{
Plots mean CV error with error bars across the log of \eqn{\lambda} sequence and
marks the optimal \code{lambda.min} and \code{lambda.1se}.
}
